refactor:     - 文件入口修改     - 分离部分结构(完成度：20%)
