+++ .env
@@ -1,0 +1,10 @@
+VITE_DEEPSEEK_API_KEY = your_deepseek_api_key
+
+
+# Love
+VITE_LOVE_TITLE = ❤️ 玺朽的砂糖 ❤️
+VITE_LOVE_ID = 超级无敌腹黑的泄兔子
+VITE_LOVE_VIEW_SAY = 你是我生命中的一部分，是我最宝贵的记忆，是我最珍贵的礼物。
+VITE_LOVE_IMG_SAY = 我会一直陪伴你，无论你走到哪里，无论你遇到什么，无论你做什么，我都会一直陪伴你。
+VITE_LOVE_BGM_SAY = 一直，一直永远在一起！
+
