+++ .gitignore
@@ -7,9 +7,14 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
+.vscode
+
+# node.js
 node_modules
 dist
 dist-ssr
+yarn.lock
+package-lock.json
 *.local
 
 # Editor directories and files
