+++ src/index.js
@@ -1,0 +1,1 @@
+[master版本内容]
