+++ src/router/index.js
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from '../views/HomeView.vue'
+import MainView from '../views/MainView.vue'
 import LoveView from '../views/LoveView.vue'
 
 export default createRouter({
@@ -8,7 +8,7 @@ export default createRouter({
     {
       path: '/',
       name: 'home',
-      component: HomeView
+      component: MainView
     },
     {
       path: '/love',
